+++ src/main/java/argonet/board/controller/CommentController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.Objects;
 
 @Controller
 @RequiredArgsConstructor
@@ -30,13 +31,19 @@ public class CommentController {
     }
 
     @PostMapping("/comment/{comment_id}/delete")
-    public String deleteComment(@PathVariable(value = "comment_id") Long comment_id, @PathVariable(value = "id") Long id) throws Exception {
+    public String deleteComment(@AuthenticationPrincipal Member member, @PathVariable(value = "comment_id") Long comment_id, @PathVariable(value = "id") Long id) throws Exception {
+        if(!Objects.equals(commentService.findById(id).getMemberId(), member.getId())){
+            return "redirect:/";
+        }
         commentService.remove(comment_id);
         return "redirect:/board/" + id;
     }
 
     @PostMapping("/comment/{comment_id}/modify")
-    public String updateComment(@PathVariable(value = "comment_id") Long comment_id, @PathVariable(value = "id") Long id, HttpServletRequest request) throws Exception {
+    public String updateComment(@AuthenticationPrincipal Member member, @PathVariable(value = "comment_id") Long comment_id, @PathVariable(value = "id") Long id, HttpServletRequest request) throws Exception {
+        if(!Objects.equals(commentService.findById(id).getMemberId(), member.getId())){
+            return "redirect:/";
+        }
         CommentRequest comment = new CommentRequest();
         comment.setId(comment_id);
         comment.setDescription(request.getParameter("description"));
